initial testing complete on splitting of files
